Refactor the request queue to store the args sent to request instead of the request context, and replay the full "request" method call
This ensures that the socket context (including all options, e.g. global headers) is fully formed by the time the request is made.  Before, the queue stored a static `requestCtx` object that the queue runner just called `_emitFrom` with.  But since `request` was being called before `_connect` (where options like `headers` are first added to the socket instance), the saved `requestCtx` object could be missing some desired user options.  Now, those options are added immediately before calling `_emitFrom`, after the socket has connected.  The one caveat is that if global options are changed between when the request is queued and when it is sent, it is the _changed_ options that will be used, not those that were present when the request was first queued.  This seems acceptable because 1) the global options are not generally intended to be changed, 2) it's not clear that using the "at-time-of-queueing" option values is a better practice and 3) this change reflects a real-world use case (setting `io.sails.headers`), and no counter-use-case currently exists, since the `request` method doesn't use any other global options besides `headers` at the moment.