Fixing commit 8393d87
On commit 8393d87, condition (this.statusCode === 0) line 539 will always be false as this.statusCode is set to 200 by default. Same for this.body line 541 that is an empty object by default. Then the error message was always set to 'The socket request failed.'.

Following code set a more explicit message.