Make shallow clones of `responseCbs` and `requestCtxs` arrays before using in `disconnect` event.  Also add underscores to...underscore that they are private vars.
Note that since these could still be accidentally modified at any time by code accessing the SailsSocket object.  This is JavaScript, after all.  We could pull the vars out of the SailsSocket entirely, but that would be a bigger task that we'd possibly want to apply to many other "private" vars as well.  Since sails.io.js supports the creation of multiple socket connections, we'd have to keep track of all the private vars in a hash.